transformed all the newRatingForm in island to use useUser, since ProductReviews is not an island anymore
